Harden graphics runtime edge cases
Tighten JPEG and PNG parsing around marker termination, EXIF orientation, tRNS validation, scanline filters, and IDCT overflow containment.

Restrict FBX texture path loading, validate world streaming and scene hierarchy growth, and harden material, collider, and physics allocation paths.

Grow keyboard input buffers, rescan Linux pads for hotplug and disconnect handling, and keep KeyChord counters and allocation paths from overflowing.

Validate AVI and GIF media budgets and frame decode paths, and keep video seeking/display state in sync at frame boundaries.

Add graphics cppcheck suppressions plus regression coverage for invalid PNG tRNS chunks, Adam7 filters, and file-dialog path-list escaping.

+++ src/runtime/graphics/3d/world/rt_particles3d.c
@@ -785,10 +785,10 @@ static void spawn_particle(rt_particles3d *ps) {
     /* Velocity */
     float dir[3];
     random_cone_dir(ps, ps->emit_dir, ps->emit_spread, dir);
-    float speed = rand_range(ps, ps->speed_min, ps->speed_max);
-    p->vel[0] = dir[0] * speed;
-    p->vel[1] = dir[1] * speed;
-    p->vel[2] = dir[2] * speed;
+    double speed = (double)rand_range(ps, ps->speed_min, ps->speed_max);
+    p->vel[0] = (double)dir[0] * speed;
+    p->vel[1] = (double)dir[1] * speed;
+    p->vel[2] = (double)dir[2] * speed;
 
     /* Life */
     p->max_life = rand_range(ps, ps->life_min, ps->life_max);
